Setup auth module user entity user repository

+++ server/src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum'
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -13,32 +14,18 @@ export class TasksService {
         private TaskRepository: TaskRepository,
     ) {}
     
-    getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-      return this.TaskRepository.getTask(filterDto); 
+    getTasks(
+        filterDto: GetTasksFilterDto,
+        user: User,
+    ): Promise<Task[]> {
+        return this.TaskRepository.getTask(filterDto, user); 
     }
-    // getAllTasks(): Task[] {
-    //     return this.tasks;
-    // }
 
-    // getTasksWithFilers(filterDto: GetTasksFilterDto): Task[] {
-    //     const { status, search } = filterDto;
-
-    //     let tasks = this.getAllTasks();
-
-    //     if ( status ) {
-    //         tasks = tasks.filter(task => task.status === status);
-    //     }
-
-    //     if ( search ) {
-    //         tasks = tasks.filter( task => 
-    //             task.title.includes(search) ||
-    //             task.description.includes(search),  
-    //         );
-    //     }
-    //     return tasks;
-    // }
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.TaskRepository.findOne(id);
+    async getTaskById(
+        id: number,
+        user: User,
+    ): Promise<Task> {
+        const found = await this.TaskRepository.findOne({ where: { id, userId: user.id } });
 
         if(!found) {
             throw new NotFoundException(`Task with ID ${id} not found`);
@@ -47,20 +34,30 @@ export class TasksService {
         return found;
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.TaskRepository.createTask(createTaskDto); 
+    async createTask(
+        createTaskDto: CreateTaskDto,
+        user: User
+    ): Promise<Task> {
+        return this.TaskRepository.createTask(createTaskDto, user); 
     }
-   
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.TaskRepository.delete(id);
+
+    async deleteTask(
+        id: number,
+        user: User,
+    ): Promise<void> {
+        const result = await this.TaskRepository.delete( { id, userId: user.id } );
 
         if(result.affected === 0) {
             throw new NotFoundException(`Task with ID ${id} not found`);
         }
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus) {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number, 
+        status: TaskStatus,
+        user: User
+    ) {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
